Resolves #40, create an access override module.
The module is utilizing the hook_node_access to allow users with special
Expert role to edit the h5p_content nodes for which the topic field or
its parents match at least one of the subjects from the user profile.
Module has a settings page that allows to selct the role and both fields
in question. All the items should be created manually and only then
selected and stored within the corresponding variables. Module would
also hide the field that holds the information about subjects from any
user that is not an administrator (portal management role check is
used instead or plain role check).