Added initial implementation of analytics module.
The implementation is a simple grab anything that is provided and send
to the xAPI LRS endpoint. It provides a configuration page for defining
LRS endpoint URL and credentials for Basic Auth. The implementation is
using as much functionality Drupal can offer as possible. There are two
logics of capturing statements: one for usage within the nomral pages
with all the benefits of Drupal JS framework and the other one is for
H5P embed that solely relies on H5P capabilities. These two scripts are
loaded in different contexts and would never be loaded at the same time.
The code is rather similar. The AJAX endpoint would respond with certain
code if something goes wrong, it would even try to proxy the response
code that was returned from the LRS endpoint. Request context
information for each failed request would be captured and logged by the
watchdog. This would still need some testing before going live.