References #106, added token protection and reCaptcha.
Using the same token as H5P itself for the setFinished endpoint. The
reason is that it is really tricky to pass any data to these scripts. It
gets even more tricky in case of EMBED view. The token should be good
enough anyway. reCaptcha is protecting the action which creats a new
challenge, this one is sending an e-mail notification and needs extra
protection. Site Key extraction is pretty trickey, to say the least.
Code would still need to checks and polish in a couple of places.