Changed analytics inner workings.
Added 'timestamp' to the statements being sent. This way the time of the
statement would still be correct even if it gets to the LRS a bit later.
Statements are no longer sent straight to LRS but collected locally,
into the queue. Then processed into batches and after that the batches
are being sent. A few more cases (network unavailable or http query
failure) could be handled and process retried, but that could be
improved upon later.