Better logic or statement timestamp.
The "within a browser" approach worked well enough most of the time, but
in some cases the dates were off by minutes or even hours. As of late,
there was one being one month in the future. This is not a reliable
solution. Had to move the timestamp logic to the backend side. That way
something could go wrong only when the server datetime value is off.