References #129, implemented new logic to fix the issue.
New logic is checking if the score and maximum score could be obtained
from the instance. If that is so and the score has not been submitted
yet, then the current one would be submitted when challenge end button
would be pressed. New logic also changed the contents of the cookie,
adding the score and time of the last time the challenge score was
stored. It does require a bit more testing and some of the texts would
probably need to be changed to reflect the new logic.