References #150, created a supplemental module
This new module would use google_tag as a base and would add the Google
Tag Manager to the external embed view. It uses the h5p_scripts hook to
add the script file generated by the base module. README file also has
instructions on how to change the codebase of h5p module in order to
allow HTML part to be added to the document body.